feat: Use @istanbuljs/schema to pull defaults (#485)
Fixes #460

BREAKING CHANGE: The defaults for `autoWrap`, `preserveComments`,
`esModules` and `produceSourceMap` are now true.  This applies only to
the stand-alone instrumenter, the visitor does not use these options.

BREAKING CHANGE: The `flow` and `jsx` parser plugins are no longer
enabled by default.  This applies only to the stand-alone instrumenter,
the visitor does not use this option.

BREAKING CHANGE: The `plugins` option of the stand-alone instrumenter
has been renamed to `parserPlugins` to match nyc.